requirements.txt
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
google==3.0.0
google-auth==2.22.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-reflection==1.43.0
grpcio-tools==1.56.2
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
Jinja2==3.1.2
jsonschema==4.10.0
kubernetes==24.2.0
kubernetes-validate==1.24.0
MarkupSafe==2.1.1
multidict==6.0.2
mypy-protobuf==3.5.0
oauthlib==3.2.0
# orchestrator-utils==0.2
pandas==2.0.0
protobuf==4.23.4
protobuf-distutils==1.0
protobuf3==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
types-protobuf==4.23.0.3
urllib3==1.26.12
websocket-client==1.4.1
